fix: try to fail faster
This may not be faster in benches but when parsing real
data from the wild it will found faster that 112.3 MiB
is not a date.

All date formats contains a time and surprisingly they
all contains a time and such time always separates hours
and minutes with a colon (':') character.